Pass the targets instead of the BUILD file to `pants--complete-read'
The purpose of this function is to present a list of choices to the end
user. It should not worry about getting the list of targets, but instead
let the caller pass it, as we do for the prompt.

We now have a new function to fetch the targets from the BUILD file.

The various interactive functions have now a smaller footprint, and most
just need to call the `pants--complete-read' function with a prompt, a
list of choices, and an action.