Try to do the right thing for the default-directory.
When getting the list of targets, or running an action, the value for
the default directory should not be changed. Until now, if you were to
open a file, then get the list of targets, then open a new file,
you'll be prompted for a new file with a starting directory set by the
value of `pants--get-source-tree`.

We now set a value for `default-directory` as a local variable when
running an action.